Updated Guide Facility

+++ app/Http/Controllers/Api/V1/GuideServiceFacilityController.php
@@ -1,7 +1,8 @@
 <?php
 
-namespace App\Http\Controllers\Api\V1;
+namespace App\Http\Controllers\Api\v1;
 
+use App\Models\Service;
 use App\Models\Facility;
 use Illuminate\Http\Request;
 use App\Http\Controllers\Controller;
@@ -10,23 +11,29 @@ use Illuminate\Support\Facades\Cache;
 use Illuminate\Support\Facades\Response;
 use Illuminate\Support\Facades\Validator;
 use App\Http\Requests\Api\V1\StoreFacilityRequest;
-use App\Models\Service;
 
-class ServiceFacilityController extends Controller
+class GuideServiceFacilityController extends Controller
 {
-
     /**
      * Store a newly created resource in storage.
      */
     public function register(StoreFacilityRequest $request)
     {
         try {
-            $agency = $request->user("agency");
-            $service = Service::where("agency_id", $agency->id)->where("id", $request->service_id)->firstOrFail();
+            $guide = $request->user("guide");
+            if ($guide->agency_id != 0) {
+                return Response::json([
+                    'success'   => false,
+                    'status'    => HTTP::HTTP_FORBIDDEN,
+                    'message'   => "You are not allowed to create facility.",
+                ],  HTTP::HTTP_FORBIDDEN); // HTTP::HTTP_OK
+            }
+
+            $service = Service::where("guide_id", $guide->id)->where("id", $request->service_id)->firstOrFail();
 
             $facility = new Facility();
             $facility->title  = $request->title;
-            $facility->agency_id  = $agency->id;
+            $facility->guide_id  = $guide->id;
             $facility->service_id  = $service->id;
             $facility->icon_id  = $request->icon_id;
             $facility->description  = $request->description;
@@ -48,8 +55,19 @@ class ServiceFacilityController extends Controller
     public function destroy(Request $request)
     {
         try {
-            $agency = $request->user("agency");
-            $facility = Facility::where("agency_id", $agency->id)->where("id", $request->facility)->firstOrFail();
+            $guide = $request->user("guide");
+
+            if ($guide->agency_id != 0) {
+                return Response::json([
+                    'success'   => false,
+                    'status'    => HTTP::HTTP_FORBIDDEN,
+                    'message'   => "You are not allowed to delete facility.",
+                ],  HTTP::HTTP_FORBIDDEN); // HTTP::HTTP_OK
+            }
+
+
+
+            $facility = Facility::where("guide_id", $guide->id)->where("id", $request->facility)->firstOrFail();
             $facility->delete();
 
             return Response::json([
